ActiScore Complete/ActiScore Latest/requirements.txt
# Core dependencies
# Web Framework
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-SocketIO==5.1.1
Flask-RESTx==1.0.3
gunicorn==20.1.0

# Database
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
pymongo==3.12.0

# ML & Data Processing
tensorflow==2.8.0
torch==1.10.0
torchvision==0.11.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==1.0
librosa==0.8.1
opencv-python==4.5.3.56
dlib==19.22.1
face-recognition==1.3.0
transformers==4.11.3
matplotlib==3.4.3
seaborn==0.11.2
plotly==5.3.1

# Audio Processing
sounddevice==0.4.2
pydub==0.25.1
pyaudio==0.2.11

# Utilities
python-dotenv==0.19.1
Pillow==8.3.2
tqdm==4.62.3